pyproject.toml
[tool.poetry]
name = "arbiter"
version = "0.1.2"
description = ""
authors = ["kc-softmax"]
readme = "README.md"

[tool.poetry.scripts]
arbiter = "arbiter.cli.main:app"

[tool.poetry.dependencies]
python = "<4.0,>=3.10"
prisma = "^0.13.1"
fastapi = "^0.97.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
sqlmodel = "^0.0.8"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^3.1.7"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
pydantic = {extras = ["dotenv"], version = "^1.10.9"}
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"
greenlet = "^3.0.3"
pytest = "^7.4.0"
jinja2 = "^3.1.2"
asyncpg = "^0.28.0"
httpx = "^0.25.0"
typer = {extras = ["all"], version = "^0.9.0"}
sqlalchemy-json = "^0.7.0"
mako = "^1.3.0"
alembic = "^1.13.1"
redis = "^5.0.1"
py-mon = "^2.0.5"
aiortc = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"